+++ src/main/java/org/app/manager/library/service/BookService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface BookService {
     List<Book> findAllBooks();
-    Book findBookById(int id);
+    Book findBookById(Long id);
     Book saveBook(Book book);
-    void deleteBookById(int id);
+    void deleteBookById(Long id);
 }
