+++ src/main/java/org/app/manager/library/service/MemberService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface MemberService {
     List<Member> findAllMembers();
-    Member findMemberById(int id);
+    Member findMemberById(Long id);
     Member saveMember(Member member);
-    void deleteMemberById(int id);
+    void deleteMemberById(Long id);
 }
